requirements.txt
pygame>=2.6.1
pillow>=12.2.0
moderngl>=5.12.0
numpy>=2.0.0
